fix: 나머지 빌드 오류 수정
- RimAIRunMetrics.cs: EndReason enum 전체 경로 지정 (GA.EndReason.Unknown)
- CinematicCameraHelper.cs: CellsVisible -> rootSize로 변경
- ConstructionSubsystem.cs: RimAIPlayStyle enum 네임스페이스 추가
- RimAISettingsUI.cs: RimWorld using 추가 (MessageTypeDefOf)
- WorkGiverPatches.cs: Traverse로 private pawn 필드 접근, Cooking defName 비교
- MedicalAnalyzer.cs: TendPatientUrgent 제거
- IRimAISubsystem.cs: Enabled 프로퍼티 backing field 추가
- BuildingPlacer.cs: isFrame -> IsFrame (대문자)
- ProductionAnalyzer.cs: StatBaseDef 대신 GetStatValue 사용
- CombatDefenseSubsystem.cs: Average() 결과 명시적 float 캐스트
- ConstructionAnalyzer.cs, BlueprintExecutor.cs: Blocks* DefDatabase 조회
- DebugHUD.cs: RimWorld using 추가 (GenDate)

+++ Source/RimAI/Combat/CombatDefenseSubsystem.cs
@@ -288,16 +288,16 @@ namespace RimAI.Combat
                 var colonists = map.mapPawns.FreeColonistsSpawned;
                 if (colonists.Any())
                 {
-                    float avgX = colonists.Average(p => p.Position.x);
-                    float avgZ = colonists.Average(p => p.Position.z);
+                    float avgX = (float)colonists.Average(p => p.Position.x);
+                    float avgZ = (float)colonists.Average(p => p.Position.z);
                     return new IntVec3((int)avgX, 0, (int)avgZ);
                 }
 
                 return map.Center;
             }
 
-            float x = buildings.Average(b => b.Position.x);
-            float z = buildings.Average(b => b.Position.z);
+            float x = (float)buildings.Average(b => b.Position.x);
+            float z = (float)buildings.Average(b => b.Position.z);
             return new IntVec3((int)x, 0, (int)z);
         }
